+++ src/simulation/onboardDataHandling/transmitter/simpleTransmitter.cpp
@@ -130,7 +130,7 @@ int SimpleTransmitter::getMaxIndex() {
     for (long unsigned int i = 0; i < this->storageUnitMsgs.back().storedData.size(); i++) {
         if (this->storageUnitMsgs.back().storedData[i] > maxVal) {
             maxVal = this->storageUnitMsgs.back().storedData[i];
-            maxIndex = i;
+            maxIndex = (int) i;
         }
     }
